Move pool with all the unsafe code into own mod
